requirements.txt
langchain
python-dotenv
streamlit
python-docx
openai
